update assemblyscript & contracts
This uses the contracts & assemblyscript settings from
https://github.com/nearprotocol/guest-book

It also updates dependencies to match, and streamlines the `test`
command to support easier testing from within an editor (adding
`testEnvironment` from the CLI command to the config)

Note that it does NOT update the frontend to match this new contract!
This breaks the app at present.