refactor: get rid of extraneous prop to App.js
Given that the App component already receives `wallet` and `wallet` can
fetch `account()._state`, it seems redundant to also pass `account`

It also appears that other internal state in near-api-js is not cleaned
up with the call to `signOut`, so the best way to reset to a clean state
after sign out is probably to refresh the page (`window.location =
'/'`). This also means that there is no reason to add `accountId` to
component state.